Add Initial Data Loading and Image Preview
In this commit, the code is implemented to load the MNIST dataset from CSV files and previewed a few records to get an initial look at the data. The images have been displayed using Matplotlib, allowing us to inspect the handwritten digits and their corresponding labels. This marks the start of our digit recognition project.